submit H.cpp to 'H - Stamps 3' (xmascon19) [C++14 (GCC 5.4.1)]

+++ 2019/1224_xmascon19/H.cpp
@@ -357,7 +357,7 @@ private:
       }
       for (auto k = 0; k < p; k++)
       {
-        if (E[k].size() < X / 2)
+        if (static_cast<int>(E[k].size()) < X / 2)
         {
           continue;
         }
